格式整改

+++ interfaces/kits/js/napi/src/get_work_status.cpp
@@ -20,10 +20,10 @@
 
 namespace OHOS {
 namespace WorkScheduler {
-const int32_t WORK_ID_INDEX = 0;
-const int32_t CALLBACK_INDEX = 1;
-const int32_t GET_WORK_STATUS_MIN_PARAMS = 1;
-const int32_t GET_WORK_STATUS_MAX_PARAMS = 2;
+const size_t WORK_ID_INDEX = 0;
+const size_t CALLBACK_INDEX = 1;
+const size_t GET_WORK_STATUS_MIN_PARAMS = 1;
+const size_t GET_WORK_STATUS_MAX_PARAMS = 2;
 
 struct GetWorkStatusParamsInfo {
     int32_t workId;
@@ -40,7 +40,7 @@ struct AsyncCallbackInfoGetWorkStatus {
 
 napi_value ParseParameters(const napi_env &env, const napi_callback_info &info, GetWorkStatusParamsInfo &params)
 {
-    int32_t argc = GET_WORK_STATUS_MAX_PARAMS;
+    size_t argc = GET_WORK_STATUS_MAX_PARAMS;
     napi_value argv[GET_WORK_STATUS_MAX_PARAMS] = {nullptr};
     NAPI_CALL(env, napi_get_cb_info(env, info, &argc, argv, NULL, NULL));
     bool paramsCheck = (argc == GET_WORK_STATUS_MIN_PARAMS) || (argc == GET_WORK_STATUS_MAX_PARAMS);
